move test_db_defrag to shaky tests
